requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via backup_script (setup.py)
fire==0.6.0
    # via backup_script (setup.py)
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via backup_script (setup.py)
loguru==0.7.2
    # via backup_script (setup.py)
six==1.16.0
    # via fire
smmap==5.0.1
    # via gitdb
tabulate==0.9.0
    # via backup_script (setup.py)
termcolor==2.4.0
    # via fire
tqdm==4.66.4
    # via backup_script (setup.py)
